Distribution pages: prepend file-specific icon to title. Fixes #1903
